requirements.txt
pytest==6.2.4
Appium-Python-Client~=2.2.0
selenium~=4.17.2
uiautomator2~=3.2.3
allure-pytest==2.13.2
allure-python-commons~=2.13.2
pillow~=10.4.0
weditor~=0.7.4.dev1
tornado~=6.4.1
logzero~=1.7.0
adbutils~=2.8.0
facebook-wda~=1.4.9
requests~=2.32.3
six~=1.16.0
setuptools~=74.1.2